Add CMake configuration that interoperates with LLVM's CMake build system.
This CMake configuration allows cheetah to be integrated with
opencilk-project build system.  This integration makes it simpler to
configure the compilation of OpenCilk, and it makes it easy to rebuild
all relevant OpenCilk components during the incremental development
process.  Furthermore, compilation of OpenCilk components can be
tailored to the presence or absence of cheetah in a safer and more
robust way.

This configuration still allows cheetah to be built as a standalone
project.  In addition, such a standalone build can be configured using
standard CMake configuration variables.