Modify and optimize the compiler-runtime ABI:
- Include more functionality in the runtime-ABI bitcode file,
  including the implementation of cilk_sync and the call to
  __builtin_setjmp in the implementation of a spawn.

- Use different enter/leave functions for a spawning function vs. a
  spawn-helper function.

- Inline calls to __cilkrts_get_tls_worker().

- Optimize accesses to __cilkrts_stack_frame fields.

Modify the logic for starting a cilkified region so that the original
thread of the program (i.e., the the boss thread) begins executing the
cilkified region directly, impersonating a worker in the process.  The
boss thread continues impersonating a worker until it either completes
the cilkified region or would perform a random steal.

Add support for using futexes for worker coordination on Linux, rather
than using pthread condition variables.