Replace logic for putting workers to sleep after they have failed many steal attempts consecutively.
The new system keeps track of three categories of workers: active,
deprived, and disengaged.

- An active worker is either executing work or has not failed too many
steal attempts consecutively.

- A deprived worker has failed a minimum number of steal attempts
  consecutively, but is still performing steal attempts.

- A disengaged worker is not performing steal attempts, but instead is
  waiting on a synchronization variable (e.g., a futex or condition
  variable) to be reengaged.

Conceptually, the new system tries to keep the number of deprived
workers no larger than the number of active workers.  Other deprived
workers are disengaged.  At a high level, the runtime system
implements the following protocol to achieve this goal:

- After a worker fails enough consecutive steal attempts, it becomes
  deprived.

- When a deprived worker fails sufficiently many consecutive steal
  attempts, it checks the runtime system to determine if there are too
  many deprived workers.  If so, it disengages.

- When a deprived worker successfully steals some computation, it
  becomes active and it reengages 2 disengaged workers.

The count of disengaged and deprived workers is maintained using
atomic operations, to ensure that all workers observe a consistent
state when executing the above protocol.

Disengaged workers wait on a variable similar to a semaphore.  To
reengage some disengaged workers, this variable is increased.  A
disengaged worker waits as long as this variable is 0.  When the
variable becomes positive, it attempts to atomically decrement the
variable.  If the disengaged worker succeeds, it reengages in work
stealing; otherwise it continues waiting.

To facilitate synchronization between workers disengaging themselves
and workers trying to reengage other workers, the variable disengaged
workers wait on may be increased up to nworkers/2, even if there are
no disengaged workers.

The system restricts workers to steal only from engaged workers by
maintaining a map of indices to workers.  To perform a random steal, a
worker chooses a random index from [0,e), where e is the number of
engaged workers.  This index is mapped to an engaged worker using the
index_to_worker map in the global state.  This map is maintained as
workers disengage and reengage.  Although the set of engaged workers
includes deprived workers, which have no work available to steal,
because the number of deprived workers is generally kept no larger
than the number of active workers, this strategy still maintains a
constant probability of stealing from an active worker.

To reduce the overhead of accessing the index_to_worker map, thieves
will use their own local copy of the index_to_worker map, which they
referesh every few steal attempts.  Even though a thief may therefore
use a stale copy of this map for some steal attempts, those steal
attempts are nevertheless safe.