[fiber] Prototype integration with ASan, to inform ASan whenever a Cilk worker thread changes stacks.
This change fixes many issues with using ASan to check for memory
errors in Cilk programs.  Some issues remain, however.  For example,
ASan may report a spurious memory leak if the Cilk program creates a
hypermap that is freed when the runtime terminates.  In this case, the
ASan library may be unloaded before the Cilk runtime deallocates the
hypermap, causing ASan to see the hypermap as allocated but never
freed.