[cilk2c_inlined] For all systems, get tls_worker from the stack frame, rather than by calling __cilkrts_get_tls_worker(), for ABI functions that may run after an arbitrary function call in the user program.
This change works around a code-generation bug in the compiler for
accesses to a worker's TLS.  The change also removes previous
special-casing of the code to look up tls_worker on different systems.
But this change also means we cannot yet remove the worker pointer
from __cilkrts_stack_frame.

This commit addresses issue OpenCilk/opencilk-project#65.